requirements.txt
typer
python-dotenv
RPi.GPIO
